Parser changes for Drupal 6 compatibility done over the weekend:  - we use dots in file names so we should use pathinfo() instead of plain old explode()  - added JS parsing based on core regex code with custom error reporting, now parsing .js too  - watchdog() parameters changed, now t() should not be used  - the menu hook changed, we identify title and description values if literals (otherwise throw an error)  - format_plural() calls now accept arguments, so we allow a comma after the plural string  - .info file format changed, so we need to use drupal_parse_info_file() for parsing
Note that I did not implement a code extract for the error message due to some laziness
(it requires watching until the line end, for which we have no code yet, and it needs
to take nesting into account), otherwise new functionality signals errors as before.

(This code also helped identify some watchdog(t()) issues in Drupal 6 which I'll file an issue to fix).